- Consolidate PivotQuery and AggregationQuery story. With OnBeforeQuery, DistinctValueDimension can be an ordinary AggregationDimension, which substantially simplifies everything.
